requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.6
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
Markdown==3.8.2
MarkupSafe==3.0.2
multidict==6.6.4
ollama==0.5.3
packaging==25.0
playwright==1.54.0
propcache==0.3.2
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.4.0
regex==2025.9.1
requests==2.32.4
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.7
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
zipp==3.23.0
